Find completion start position using swift-syntax instead of ASCII-based implemenation
rdar://121322872
